Time: 15 ms (36.44%) | Memory: 40.1 MB (27.33%) - LeetSync

+++ 2090-number-of-ways-to-arrive-at-destination/number-of-ways-to-arrive-at-destination.cpp
@@ -16,7 +16,7 @@ public:
         pq.push({0, 0});
         // ways[0]=1;
         
-        int mod = (long int)(1e9 + 7);
+        int mod = 1e9 + 7;
         
         while(!pq.empty()){
             long int d = pq.top().first;
